Split FileData into 3 classes
Also add builder for resource properties
Also minor cleanups

Test: Existing
